requirements.txt
apache-airflow-providers-http
apache-airflow-providers-postgres
psycopg2-binary